+++ packages/devui-vue/devui/cascader/components/cascader-item/index.scss
@@ -4,10 +4,11 @@
 .#{$devui-prefix}-cascader {
   &__li {
     &.dropdown-item {
-      height: 32px;
+      height: 36px;
       padding: 8px 12px;
       color: $devui-text;
       cursor: pointer;
+      border-radius: $devui-border-radius;
       @include flex(flex-start);
 
       margin-top: 4px;
@@ -53,6 +54,8 @@
         margin-right: 4px;
         font-size: $devui-font-size-icon;
         color: $devui-text;
+        line-height: 16px;
+        height: 16px;
 
         &.disabled {
           color: $devui-disabled-text !important;
